@@ -9,14 +9,27 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "psr/log": "^1.0",
-        "wyrihaximus/react-child-process-messenger": "^2.7"
+        "php": "^7.4",
+        "psr/log": "^1.1",
+        "react/event-loop": "^1.1",
+        "react/promise": "^2.8",
+        "wyrihaximus/react-child-process-messenger": "^3"
     },
     "require-dev": {
-        "api-clients/cs-fixer-config": "^1.1",
-        "api-clients/test-utilities": "^5.4",
-        "monolog/monolog": "^1.22 || ^2.0"
+        "monolog/monolog": "^2.2",
+        "wyrihaximus/async-test-utilities": "^3.4.18"
+    },
+    "config": {
+        "platform": {
+            "php": "7.4.7"
+        },
+        "sort-packages": true
+    },
+    "extra": {
+        "unused": [
+            "php",
+            "react/promise-stream"
+        ]
     },
     "autoload": {
         "psr-4": {
@@ -26,66 +39,14 @@
     "autoload-dev": {
         "psr-4": {
             "WyriHaximus\\React\\Tests\\ChildProcess\\PSR3\\": "tests/"
-        },
-        "files": ["src/functions_include.php"]
-    },
-    "config": {
-        "sort-packages": true,
-        "platform": {
-            "php": "7.2"
         }
     },
     "scripts": {
-        "ensure-installed": "composer install --ansi -n -q",
-        "cs": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --dry-run --diff --verbose --allow-risky=yes --show-progress=estimating"
+        "post-install-cmd": [
+            "composer normalize"
         ],
-        "cs-fix": [
-            "@ensure-installed",
-            "php-cs-fixer fix --config=.php_cs --ansi --verbose --allow-risky=yes --show-progress=estimating"
-        ],
-        "unit": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist"
-        ],
-        "unit-coverage": [
-            "@ensure-installed",
-            "phpunit --colors=always -c phpunit.xml.dist --coverage-text --coverage-html covHtml --coverage-clover ./build/logs/clover.xml"
-        ],
-        "lint-php": [
-            "@ensure-installed",
-            "parallel-lint --exclude vendor ."
-        ],
-        "qa-all": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-all-coverage": [
-            "@lint-php",
-            "@cs",
-            "@unit-coverage"
-        ],
-        "qa-windows": [
-            "@lint-php",
-            "@cs",
-            "@unit"
-        ],
-        "qa-ci": [
-            "@unit"
-        ],
-        "qa-ci-extended": [
-            "@qa-all-coverage"
-        ],
-        "qa-ci-windows": [
-            "@qa-windows"
-        ],
-        "qa-contrib": [
-            "@qa-all"
-        ],
-        "ci-coverage": [
-            "if [ -f ./build/logs/clover.xml ]; then wget https://scrutinizer-ci.com/ocular.phar && php ocular.phar code-coverage:upload --format=php-clover ./build/logs/clover.xml; fi"
+        "post-update-cmd": [
+            "composer normalize"
         ]
     }
 }
