@@ -1,69 +1,60 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace WyriHaximus\React\ChildProcess\PSR3;
 
-use InvalidArgumentException;
 use Psr\Log\LoggerInterface;
 use React\EventLoop\LoopInterface;
 use React\Promise\PromiseInterface;
-use function React\Promise\reject;
-use function React\Promise\resolve;
 use Throwable;
 use WyriHaximus\React\ChildProcess\Messenger\ChildInterface;
 use WyriHaximus\React\ChildProcess\Messenger\Messages\Payload;
-use WyriHaximus\React\ChildProcess\Messenger\Messenger;
+use WyriHaximus\React\ChildProcess\Messenger\MessengerInterface;
+
+use function React\Promise\reject;
+use function React\Promise\resolve;
 
 final class ChildProcess implements ChildInterface
 {
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
+    private LoggerInterface $logger;
 
-    private function __construct(Messenger $messenger, LoopInterface $loop)
+    private function __construct(MessengerInterface $messenger, LoggerInterface $logger)
     {
-        $messenger->registerRpc('logger.setup', function (Payload $payload) {
-            return $this->setup($payload);
-        });
-        $messenger->registerRpc('logger.log', function (Payload $payload) {
+        $this->logger = $logger;
+        $messenger->registerRpc('logger.log', function (Payload $payload): PromiseInterface {
             return $this->log($payload);
         });
     }
 
-    private function setup(Payload $payload): PromiseInterface
+    public static function create(MessengerInterface $messenger, LoopInterface $loop): void
     {
-        try {
-            $this->logger = $payload['factory']();
-            if (!($this->logger instanceof LoggerInterface)) {
-                throw new InvalidArgumentException('Passed logger isn\'t a PSR-3 logger');
-            }
+        $loop->futureTick(static function (): void {
+        });
+        $messenger->registerRpc('logger.setup', static function (Payload $payload) use ($messenger): PromiseInterface {
+            /**
+             * @psalm-suppress PossiblyNullFunctionCall
+             */
+            new static($messenger, $payload['factory']());
 
-            return resolve([
-                'success' => true,
-            ]);
-        } catch (Throwable $throwable) {
-            return reject($throwable);
-        }
-    }
-
-    public static function create(Messenger $messenger, LoopInterface $loop): void
-    {
-        new static($messenger, $loop);
+            return resolve(['success' => true]);
+        });
     }
 
     private function log(Payload $payload): PromiseInterface
     {
         try {
+            /**
+             * @psalm-suppress PossiblyNullArgument
+             */
             $this->logger->log(
                 $payload['level'],
                 $payload['message'],
                 $payload['context']
             );
 
-            return resolve([
-                'success' => true,
-            ]);
-        } catch (Throwable $throwable) {
+            return resolve(['success' => true]);
+        } catch (Throwable $throwable) { /** @phpstan-ignore-line */
             return reject($throwable);
         }
     }
